refactor: ClubApplicationFormSlim 위치 dto로 변경
